+++ packages/editor/src/lib/config/defaultShapes.ts
@@ -1,10 +1,10 @@
 import { arrowShape } from '../editor/shapes/ArrowShape/arrowShape'
 import { drawShape } from '../editor/shapes/DrawShape/drawShape'
+import { geoShape } from '../editor/shapes/GeoShape/geoShape'
 import { highlightShape } from '../editor/shapes/HighlightShape/highlightShape'
 import { BookmarkShapeUtil } from '../editor/shapeutils/BookmarkShapeUtil/BookmarkShapeUtil'
 import { EmbedShapeUtil } from '../editor/shapeutils/EmbedShapeUtil/EmbedShapeUtil'
 import { FrameShapeUtil } from '../editor/shapeutils/FrameShapeUtil/FrameShapeUtil'
-import { GeoShapeUtil } from '../editor/shapeutils/GeoShapeUtil/GeoShapeUtil'
 import { GroupShapeUtil } from '../editor/shapeutils/GroupShapeUtil/GroupShapeUtil'
 import { ImageShapeUtil } from '../editor/shapeutils/ImageShapeUtil/ImageShapeUtil'
 import { LineShapeUtil } from '../editor/shapeutils/LineShapeUtil/LineShapeUtil'
@@ -44,9 +44,9 @@ export const coreShapes: Record<string, TLShapeInfo> = {
 /** @public */
 export const defaultShapes: Record<string, TLShapeInfo> = {
 	draw: drawShape,
-	geo: {
-		util: GeoShapeUtil,
-	},
+	arrow: arrowShape,
+	highlight: highlightShape,
+	geo: geoShape,
 	line: {
 		util: LineShapeUtil,
 	},
@@ -56,6 +56,4 @@ export const defaultShapes: Record<string, TLShapeInfo> = {
 	frame: {
 		util: FrameShapeUtil,
 	},
-	arrow: arrowShape,
-	highlight: highlightShape,
 }
