+++ packages/editor/src/lib/config/defaultTools.ts
@@ -1,9 +1,9 @@
 import { ArrowShapeTool } from '../editor/shapes/ArrowShape/ArrowShapeTool/ArrowShapeTool'
 import { DrawShapeTool } from '../editor/shapes/DrawShape/DrawShapeTool/DrawShapeTool'
+import { GeoShapeTool } from '../editor/shapes/GeoShape/GeoShapeTool/GeoShapeTool'
 import { HighlightShapeTool } from '../editor/shapes/HighlightShape/HighlightShapeTool/HighlightShapeTool'
 import { EraserTool } from '../editor/tools/EraserTool/EraserTool'
 import { FrameShapeTool } from '../editor/tools/FrameShapeTool/FrameShapeTool'
-import { GeoShapeTool } from '../editor/tools/GeoShapeTool/GeoShapeTool'
 import { HandTool } from '../editor/tools/HandTool/HandTool'
 import { LaserTool } from '../editor/tools/LaserTool/LaserTool'
 import { LineShapeTool } from '../editor/tools/LineShapeTool/LineShapeTool'
