+++ packages/editor/src/lib/editor/shapes/ArrowShape/arrowShape.ts
@@ -2,10 +2,11 @@ import { createShape } from '../../../config/createShape'
 import { ArrowShapeTool } from './ArrowShapeTool/ArrowShapeTool'
 import { ArrowShapeUtil } from './ArrowShapeUtil/ArrowShapeUtil'
 import { arrowShapeMigrations } from './arrowShapeMigrations'
+import { TLArrowShape } from './arrowShapeTypes'
 import { arrowShapeValidator } from './arrowShapeValidator'
 
 /** @public */
-export const arrowShape = createShape({
+export const arrowShape = createShape<TLArrowShape>({
 	util: ArrowShapeUtil,
 	tool: ArrowShapeTool,
 	migrations: arrowShapeMigrations,
