+++ packages/editor/src/lib/editor/shapes/DrawShape/drawShape.ts
@@ -2,10 +2,11 @@ import { createShape } from '../../../config/createShape'
 import { DrawShapeTool } from './DrawShapeTool/DrawShapeTool'
 import { DrawShapeUtil } from './DrawShapeUtil/DrawShapeUtil'
 import { drawShapeMigrations } from './drawShapeMigrations'
+import { TLDrawShape } from './drawShapeTypes'
 import { drawShapeValidator } from './drawShapeValidator'
 
 /** @public */
-export const drawShape = createShape({
+export const drawShape = createShape<TLDrawShape>({
 	util: DrawShapeUtil,
 	tool: DrawShapeTool,
 	migrations: drawShapeMigrations,
