+++ packages/editor/src/lib/editor/shapes/GeoShape/GeoShapeTool/GeoShapeTool.test.ts
@@ -1,4 +1,4 @@
-import { TestEditor } from '../../../test/TestEditor'
+import { TestEditor } from '../../../../test/TestEditor'
 import { GeoShapeTool } from './GeoShapeTool'
 
 let editor: TestEditor
