+++ packages/editor/src/lib/editor/shapes/GeoShape/GeoShapeTool/GeoShapeTool.ts
@@ -1,6 +1,5 @@
-import { StateNode } from '../StateNode'
-
 import { TLStyleType } from '@tldraw/tlschema'
+import { StateNode } from '../../../tools/StateNode'
 import { Idle } from './children/Idle'
 import { Pointing } from './children/Pointing'
