+++ packages/editor/src/lib/editor/shapes/GeoShape/GeoShapeTool/children/Idle.ts
@@ -1,5 +1,5 @@
-import { TLEventHandlers } from '../../../types/event-types'
-import { StateNode } from '../../StateNode'
+import { StateNode } from '../../../../tools/StateNode'
+import { TLEventHandlers } from '../../../../types/event-types'
 
 export class Idle extends StateNode {
 	static override id = 'idle'
