+++ packages/editor/src/lib/editor/shapes/GeoShape/GeoShapeTool/children/Pointing.ts
@@ -1,7 +1,8 @@
 import { Box2d, getStarBounds } from '@tldraw/primitives'
-import { TLGeoShape, createShapeId } from '@tldraw/tlschema'
-import { TLEventHandlers } from '../../../types/event-types'
-import { StateNode } from '../../StateNode'
+import { createShapeId } from '@tldraw/tlschema'
+import { StateNode } from '../../../../tools/StateNode'
+import { TLEventHandlers } from '../../../../types/event-types'
+import { TLGeoShape } from '../../geoShapeTypes'
 
 export class Pointing extends StateNode {
 	static override id = 'pointing'
