+++ packages/editor/src/lib/editor/shapes/GeoShape/GeoShapeUtil/components/DashStyleEllipse.tsx
@@ -1,12 +1,16 @@
-import { toDomPrecision } from '@tldraw/primitives'
-import { TLGeoShape, TLShapeId } from '@tldraw/tlschema'
+import { perimeterOfEllipse, toDomPrecision } from '@tldraw/primitives'
+import { TLShapeId } from '@tldraw/tlschema'
 import * as React from 'react'
-import { ShapeFill, getShapeFillSvg, getSvgWithShapeFill } from '../../shared/ShapeFill'
-import { TLExportColors } from '../../shared/TLExportColors'
-import { getPerfectDashProps } from '../../shared/getPerfectDashProps'
-import { getOvalPerimeter, getOvalSolidPath } from '../helpers'
+import {
+	ShapeFill,
+	getShapeFillSvg,
+	getSvgWithShapeFill,
+} from '../../../../shapeutils/shared/ShapeFill'
+import { TLExportColors } from '../../../../shapeutils/shared/TLExportColors'
+import { getPerfectDashProps } from '../../../../shapeutils/shared/getPerfectDashProps'
+import { TLGeoShape } from '../../geoShapeTypes'
 
-export const DashStyleOval = React.memo(function DashStyleOval({
+export const DashStyleEllipse = React.memo(function DashStyleEllipse({
 	w,
 	h,
 	strokeWidth: sw,
@@ -17,8 +21,12 @@ export const DashStyleOval = React.memo(function DashStyleOval({
 	strokeWidth: number
 	id: TLShapeId
 }) {
-	const d = getOvalSolidPath(w, h)
-	const perimeter = getOvalPerimeter(w, h)
+	const cx = w / 2
+	const cy = h / 2
+	const rx = Math.max(0, cx - sw / 2)
+	const ry = Math.max(0, cy - sw / 2)
+
+	const perimeter = perimeterOfEllipse(rx, ry)
 
 	const { strokeDasharray, strokeDashoffset } = getPerfectDashProps(
 		perimeter < 64 ? perimeter * 2 : perimeter,
@@ -26,11 +34,11 @@ export const DashStyleOval = React.memo(function DashStyleOval({
 		{
 			style: dash,
 			snap: 4,
-			start: 'outset',
-			end: 'outset',
 			closed: true,
 		}
 	)
+
+	const d = `M${cx - rx},${cy}a${rx},${ry},0,1,1,${rx * 2},0a${rx},${ry},0,1,1,-${rx * 2},0`
 
 	return (
 		<>
@@ -50,7 +58,7 @@ export const DashStyleOval = React.memo(function DashStyleOval({
 	)
 })
 
-export function DashStyleOvalSvg({
+export function DashStyleEllipseSvg({
 	w,
 	h,
 	strokeWidth: sw,
@@ -63,8 +71,12 @@ export function DashStyleOvalSvg({
 	id: TLShapeId
 	colors: TLExportColors
 }) {
-	const d = getOvalSolidPath(w, h)
-	const perimeter = getOvalPerimeter(w, h)
+	const cx = w / 2
+	const cy = h / 2
+	const rx = Math.max(0, cx - sw / 2)
+	const ry = Math.max(0, cy - sw / 2)
+
+	const perimeter = perimeterOfEllipse(rx, ry)
 
 	const { strokeDasharray, strokeDashoffset } = getPerfectDashProps(
 		perimeter < 64 ? perimeter * 2 : perimeter,
@@ -75,6 +87,8 @@ export function DashStyleOvalSvg({
 			closed: true,
 		}
 	)
+
+	const d = `M${cx - rx},${cy}a${rx},${ry},0,1,1,${rx * 2},0a${rx},${ry},0,1,1,-${rx * 2},0`
 
 	const strokeElement = document.createElementNS('http://www.w3.org/2000/svg', 'path')
 	strokeElement.setAttribute('d', d)
