+++ packages/editor/src/lib/editor/shapes/GeoShape/GeoShapeUtil/components/DrawStylePolygon.tsx
@@ -1,8 +1,12 @@
-import { getRoundedInkyPolygonPath, getRoundedPolygonPoints, VecLike } from '@tldraw/primitives'
-import { TLGeoShape } from '@tldraw/tlschema'
-import * as React from 'react'
-import { getShapeFillSvg, getSvgWithShapeFill, ShapeFill } from '../../shared/ShapeFill'
-import { TLExportColors } from '../../shared/TLExportColors'
+import { VecLike, getRoundedInkyPolygonPath, getRoundedPolygonPoints } from '@tldraw/primitives'
+import React from 'react'
+import {
+	ShapeFill,
+	getShapeFillSvg,
+	getSvgWithShapeFill,
+} from '../../../../shapeutils/shared/ShapeFill'
+import { TLExportColors } from '../../../../shapeutils/shared/TLExportColors'
+import { TLGeoShape } from '../../geoShapeTypes'
 
 export const DrawStylePolygon = React.memo(function DrawStylePolygon({
 	id,
