+++ packages/editor/src/lib/editor/shapes/GeoShape/GeoShapeUtil/helpers.ts
