+++ packages/editor/src/lib/test/commands/rotateShapes.test.ts
@@ -1,5 +1,5 @@
 import { createShapeId } from '@tldraw/tlschema'
-import { GeoShapeUtil } from '../../editor/shapeutils/GeoShapeUtil/GeoShapeUtil'
+import { GeoShapeUtil } from '../../..'
 import { TestEditor } from '../TestEditor'
 
 let editor: TestEditor
