+++ packages/editor/src/lib/test/shapeutils.test.ts
@@ -1,6 +1,6 @@
 import { createShapeId } from '@tldraw/tlschema'
+import { GeoShapeUtil } from '../..'
 import { FrameShapeUtil } from '../editor/shapeutils/FrameShapeUtil/FrameShapeUtil'
-import { GeoShapeUtil } from '../editor/shapeutils/GeoShapeUtil/GeoShapeUtil'
 import { TestEditor } from './TestEditor'
 
 let editor: TestEditor
